refactor(repo): consistency sweep across exports, naming, errors
Drops camelCase utility filenames in apps/desktop/src in favour of
kebab-case, replaces inline
"err instanceof Error ? err.message : String(err)" with the shared
formatError helper across 15 source files, sorts the @kay-am/ui barrel
and the misordered ids block in @kay-am/types. z-index ladder already
within 0..50 (no hits). No behaviour changes.

Closes #520
