+++ apps/desktop/src/__tests__/agent-row-format.test.ts
@@ -12,7 +12,7 @@ import {
   formatCost,
   formatTokens,
   shortModel,
-} from '../agentRowFormat';
+} from '../agent-row-format';
 
 describe('formatTokens', () => {
   it('renders raw count under 1k', () => {
