+++ apps/desktop/src/__tests__/store/unknown-payload.test.ts
@@ -106,7 +106,7 @@ vi.mock('../../worktree', () => ({
 
 vi.mock('../../repo', () => ({ validateGitRepo: vi.fn() }));
 
-vi.mock('../../providerPricing', () => ({
+vi.mock('../../provider-pricing', () => ({
   parseProviderPricingConfig: vi.fn(() => null),
   getCodexPriceOverride: vi.fn(() => null),
   refreshPricingTable: vi.fn(() => Promise.resolve()),
