+++ apps/desktop/src/agent-kind.ts
