+++ apps/desktop/src/agent-row-format.ts
