+++ apps/desktop/src/components/SkillsPanel.tsx
@@ -1,6 +1,7 @@
 import { useEffect, useState } from 'react';
 import { Button, Input, Textarea } from '@kay-am/ui';
 import type { Skill, SkillFrontmatter, WorkspaceId } from '@kay-am/types';
+import { formatError } from '../errors';
 import { EMPTY_ARRAY, useAppStore } from '../store';
 
 interface SkillsPanelProps {
@@ -119,7 +120,7 @@ export function SkillsPanel({ workspaceId }: SkillsPanelProps) {
       });
       setEditingSkill(null);
     } catch (err) {
-      setFormError(err instanceof Error ? err.message : String(err));
+      setFormError(formatError(err));
     } finally {
       setSaving(false);
     }
