+++ apps/desktop/src/components/WorkflowNextStepCta.tsx
@@ -3,7 +3,7 @@ import { ArrowRight } from 'lucide-react';
 import { cn } from '@kay-am/ui';
 import type { Session, Step, Workflow } from '@kay-am/types';
 import type { VerbosityLevel } from '../verbosity';
-import { AGENT_KIND_DEFAULTS, AGENT_KIND_PALETTE, inferAgentKindFromName } from '../agentKind';
+import { AGENT_KIND_DEFAULTS, AGENT_KIND_PALETTE, inferAgentKindFromName } from '../agent-kind';
 
 export interface WorkflowNextStepCtaProps {
   readonly workflow: Workflow;
