+++ apps/desktop/src/provider-pricing.test.ts
@@ -3,7 +3,7 @@ import {
   getCodexPriceOverride,
   getActivePricingTable,
   refreshPricingTable,
-} from './providerPricing';
+} from './provider-pricing';
 
 describe('getActivePricingTable', () => {
   it('returns shipped pricing by default', () => {
