+++ apps/desktop/src/store/preamble.ts
@@ -1,6 +1,6 @@
 import { serializeSlots, type SlotKey, PROVIDER_CAPABILITIES } from '@kay-am/core';
 import type { ContextSlot, TurnEvent } from '@kay-am/types';
-import { estimateTokens } from '../utils/estimateTokens';
+import { estimateTokens } from '../utils/estimate-tokens';
 
 const CONTEXT_MARKER_HINT =
   '## context handoff protocol\n' +
