+++ apps/desktop/src/store/store.audit-retry.test.ts
@@ -136,7 +136,7 @@ vi.mock('../repo', () => ({
   validateGitRepo: vi.fn(),
 }));
 
-vi.mock('../providerPricing', () => ({
+vi.mock('../provider-pricing', () => ({
   parseProviderPricingConfig: vi.fn(() => null),
   getCodexPriceOverride: vi.fn(() => null),
   refreshPricingTable: vi.fn(() => Promise.resolve()),
