+++ apps/desktop/src/store/store.end-session.test.ts
@@ -123,7 +123,7 @@ vi.mock('../repo', () => ({
   validateGitRepo: vi.fn(),
 }));
 
-vi.mock('../providerPricing', () => ({
+vi.mock('../provider-pricing', () => ({
   parseProviderPricingConfig: vi.fn(() => null),
   getCodexPriceOverride: vi.fn(() => null),
   refreshPricingTable: vi.fn(() => Promise.resolve()),
