+++ apps/desktop/src/store/store.next-actions.test.ts
@@ -72,7 +72,7 @@ vi.mock('../phases', () => ({
 
 vi.mock('../worktree', () => ({ createWorktree: vi.fn(), removeWorktree: vi.fn() }));
 vi.mock('../repo', () => ({ validateGitRepo: vi.fn() }));
-vi.mock('../providerPricing', () => ({
+vi.mock('../provider-pricing', () => ({
   parseProviderPricingConfig: vi.fn(() => null),
   getCodexPriceOverride: vi.fn(() => null),
   refreshPricingTable: vi.fn(() => Promise.resolve()),
