+++ apps/desktop/src/store/store.summarizer-queue.test.ts
@@ -81,7 +81,7 @@ vi.mock('../repo', () => ({
   validateGitRepo: vi.fn(),
 }));
 
-vi.mock('../providerPricing', () => ({
+vi.mock('../provider-pricing', () => ({
   parseProviderPricingConfig: vi.fn(() => null),
   getCodexPriceOverride: vi.fn(() => null),
   refreshPricingTable: vi.fn(() => Promise.resolve()),
