+++ apps/desktop/src/utils/display-path.ts
