+++ packages/types/src/index.ts
@@ -1,10 +1,10 @@
 export type {
   IsoDateTime,
   MessageId,
-  PermissionRequestId,
-  PermissionRuleId,
   ParallelGroupId,
   ParallelSessionId,
+  PermissionRequestId,
+  PermissionRuleId,
   ProviderRunId,
   SessionId,
   SkillId,
