+++ packages/ui/src/index.ts
@@ -3,12 +3,10 @@ export { AppShell } from './components/AppShell';
 export type { AppShellProps } from './components/AppShell';
 export { Button } from './components/Button';
 export type { ButtonProps, ButtonSize, ButtonVariant } from './components/Button';
-export { CopyButton } from './components/CopyButton';
-export type { CopyButtonProps } from './components/CopyButton';
-export { Tooltip } from './components/Tooltip';
-export type { TooltipProps, TooltipSide } from './components/Tooltip';
 export { Collapsible } from './components/Collapsible';
 export type { CollapsibleProps } from './components/Collapsible';
+export { CopyButton } from './components/CopyButton';
+export type { CopyButtonProps } from './components/CopyButton';
 export { Dialog } from './components/Dialog';
 export type { DialogProps, DialogSize } from './components/Dialog';
 export { Input } from './components/Input';
@@ -20,6 +18,8 @@ export { ScrollArea } from './components/ScrollArea';
 export type { ScrollAreaProps } from './components/ScrollArea';
 export { Select } from './components/Select';
 export type { SelectProps, SelectSize } from './components/Select';
+export { Skeleton } from './components/Skeleton';
 export { Textarea } from './components/Textarea';
 export type { TextareaProps } from './components/Textarea';
-export { Skeleton } from './components/Skeleton';
+export { Tooltip } from './components/Tooltip';
+export type { TooltipProps, TooltipSide } from './components/Tooltip';
